fix: #84 코드 리뷰 반영 — XSS 방지, SSR 전환, 접근성, Footer 링크
- JSON-LD </script> escape 추가 (XSS 방지)
- Journal 목록 페이지 SSR 전환 (SEO 개선)
- 카테고리 필터 aria-pressed + role="group" 접근성 추가
- Footer에 컬렉션/Archive/Journal 링크 추가
- Footer bg-white → bg-background 테마 토큰 수정
