+++ src/app/journal/[slug]/page.tsx
@@ -52,7 +52,7 @@ export default async function JournalDetailPage({ params }: PageProps) {
     <div className="min-h-screen">
       <script
         type="application/ld+json"
-        dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd) }}
+        dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd).replace(/</g, '\\u003c') }}
       />
 
       {/* Hero */}
